feat: complete safe student fee billing and collection workflow

+++ EduOS.Core/Entities/Finance/Discount.cs
@@ -5,11 +5,10 @@ namespace EduOS.Core.Entities.Finance
     public class Discount : BaseTenantEntity
     {
         public string Name { get; set; } = string.Empty;
-        public string Type { get; set; } = "Percentage"; // Percentage/Fixed
+        public string Type { get; set; } = "Percentage";
         public decimal Value { get; set; }
-        public int? FeeHeadId { get; set; }
+        public long? FeeHeadId { get; set; }
         public bool IsActive { get; set; } = true;
-
         public virtual FeeHead? FeeHead { get; set; }
     }
 }
